+++ library-account/src/test/java/universum/studios/android/officium/account/BaseAccountAuthenticatorTest.java
@@ -1,6 +1,6 @@
 /*
  * *************************************************************************************************
- *                                 Copyright 2017 Universum Studios
+ *                                 Copyright 2018 Universum Studios
  * *************************************************************************************************
  *                  Licensed under the Apache License, Version 2.0 (the "License")
  * -------------------------------------------------------------------------------------------------
@@ -18,24 +18,41 @@
  */
 package universum.studios.android.officium.account;
 
-import android.support.test.runner.AndroidJUnit4;
+import android.accounts.AccountAuthenticatorResponse;
+import android.content.Context;
+import android.support.annotation.NonNull;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import universum.studios.android.test.instrumented.InstrumentedTestCase;
+import universum.studios.android.test.local.RobolectricTestCase;
+
+import static org.mockito.Mockito.mock;
 
 /**
  * @author Martin Albedinsky
  */
-@RunWith(AndroidJUnit4.class)
-public final class SingleUserAccountManagerTest extends InstrumentedTestCase {
+public final class BaseAccountAuthenticatorTest extends RobolectricTestCase {
 
-	@SuppressWarnings("unused")
-	private static final String TAG = "SingleUserAccountManagerTest";
+	private static final String ACCOUNT_TYPE = "account.com";
 
-	@Test
-	public void test() {
-		// todo: implement test
+	@Test public void testInstantiation() {
+		// Act:
+		new TestAuthenticator(application);
+	}
+
+	@Test(expected = UnsupportedOperationException.class)
+	public void testEditProperties() {
+		// Arrange:
+		final AccountAuthenticatorResponse mockResponse = mock(AccountAuthenticatorResponse.class);
+		final BaseAccountAuthenticator authenticator = new TestAuthenticator(application);
+		// Act:
+		authenticator.editProperties(mockResponse, ACCOUNT_TYPE);
+	}
+
+	private static final class TestAuthenticator extends BaseAccountAuthenticator {
+
+		TestAuthenticator(@NonNull final Context context) {
+			super(context);
+		}
 	}
 }