+++ CHANGELOG.md
@@ -12,6 +12,8 @@ More **detailed changelog** for each respective version may be viewed by pressin
 - Resolved [Issue #33](https://github.com/universum-studios/android_officium/issues/33).
 - Resolved [Issue #34](https://github.com/universum-studios/android_officium/issues/34).
 - Resolved [Issue #35](https://github.com/universum-studios/android_officium/issues/35).
+- Deprecated some of delegate methods of `UserAccountManager` and `SingleUserAccountManager` which
+  only delegated to the system `AccountManager`.
 
 ## Version 1.x ##